refactor(imageCanvas): consolidate configModal inline, remove separate modal file
